Separate Bezier file in Curve, Curves and Bezier
