+++ korma/src/commonMain/kotlin/com/soywiz/korma/geom/bezier/Bezier.kt
@@ -15,102 +15,6 @@ import kotlin.math.hypot
 import kotlin.math.max
 import kotlin.math.min
 
-@JvmName("ListCurves_toCurves")
-fun List<Curves>.toCurves() = Curves(this.flatMap { it.curves })
-@JvmName("ListCurve_toCurves")
-fun List<Curve>.toCurves() = Curves(this)
-
-data class Curves(val curves: List<Curve>) : Curve {
-    data class CurveInfo(
-        val curve: Curve,
-        val start: Double,
-        val end: Double,
-        val bounds: Rectangle,
-    ) {
-        fun contains(pos: Double): Boolean = pos in start..end
-
-        val length: Double get() = end - start
-    }
-
-    val infos: List<CurveInfo> by lazy {
-        var pos = 0.0
-        curves.map {
-            val start = pos
-            pos += it.length()
-            CurveInfo(it, start, pos, it.getBounds())
-        }
-    }
-    val length: Double by lazy { infos.sumOf { it.length } }
-    private val bb = BoundsBuilder()
-
-    override fun getBounds(target: Rectangle): Rectangle {
-        bb.reset()
-        infos.fastForEach { bb.addEvenEmpty(it.bounds) }
-        return bb.getBounds(target)
-    }
-
-    override fun calc(t: Double, target: Point): Point {
-        val pos = t * length
-        val index = infos.binarySearch {
-            when {
-                it.contains(pos) -> 0
-                it.end < pos -> -1
-                else -> +1
-            }
-        }
-        val info = infos.getOrNull(index) ?: error("OUTSIDE")
-        val posInCurve = pos - info.start
-        val ratioInCurve = posInCurve / info.length
-        return info.curve.calc(ratioInCurve, target)
-    }
-
-    override fun length(steps: Int): Double = length
-}
-
-fun Curve.getPoints(count: Int = this.recommendedDivisions(), out: PointArrayList = PointArrayList()): IPointArrayList {
-    val c1 = count - 1
-    val temp = Point()
-    for (n in 0..c1) {
-        val ratio = n.toDouble() / c1
-        val point = calc(ratio, temp)
-        //println("${this::class.simpleName}: ratio: $ratio, point=$point")
-        out.add(point)
-    }
-    return out
-}
-
-interface Curve {
-    fun getBounds(target: Rectangle = Rectangle()): Rectangle
-    fun calc(t: Double, target: Point = Point()): Point
-    fun length(steps: Int = recommendedDivisions()): Double
-    fun recommendedDivisions(): Int = DEFAULT_STEPS
-
-    class Line(
-        val x0: Double = 0.0, val y0: Double = 0.0,
-        val x1: Double = 0.0, val y1: Double = 0.0,
-    ) : Bezier {
-        override fun getBounds(target: Rectangle): Rectangle =
-            target.setBounds(min(x0, x1), min(y0, y1), max(x0, x1), max(y0, y1))
-
-        override fun calc(t: Double, target: Point): Point =
-            target.setTo(t.interpolate(x0, x1), t.interpolate(y0, y1))
-                //.also { println("Line.calc[t=$t] -> $target") }
-
-        override fun length(steps: Int): Double = hypot(x1 - x0, y1 - y0)
-
-        override fun recommendedDivisions(): Int = lineRecommendedSteps(x0, y0, x1, y1)
-
-        override fun toString(): String = "Curve.Line(($x0, $y0), ($x1, $y1))"
-    }
-    companion object {
-        const val DEFAULT_STEPS = 100
-
-        fun lineRecommendedSteps(x0: Double, y0: Double, x1: Double, y1: Double): Int {
-            return DEFAULT_STEPS
-        }
-    }
-}
-
 //(x0,y0) is start point; (x1,y1),(x2,y2) is control points; (x3,y3) is end point.
 // https://pomax.github.io/bezierinfo/
 interface Bezier : Curve {
@@ -146,6 +50,10 @@ interface Bezier : Curve {
         override fun calc(t: Double, target: Point): Point = quadCalc(p0.x, p0.y, p1.x, p1.y, p2.x, p2.y, t, target)
         override fun length(steps: Int): Double = quadLength(p0.x, p0.y, p1.x, p1.y, p2.x, p2.y, steps)
         override fun recommendedDivisions(): Int = quadRecommendedSteps(p0.x, p0.y, p1.x, p1.y, p2.x, p2.y)
+
+        override fun normal(t: Double, target: Point): Point {
+            TODO("Not yet implemented")
+        }
 
         // http://fontforge.github.io/bezier.html
         fun toCubic(out: Cubic = Cubic()): Cubic {
@@ -230,6 +138,10 @@ interface Bezier : Curve {
 
         override fun calc(t: Double, target: Point): Point =
             cubicCalc(p0.x, p0.y, p1.x, p1.y, p2.x, p2.y, p3.x, p3.y, t, target)
+
+        override fun normal(t: Double, target: Point): Point {
+            TODO("Not yet implemented")
+        }
 
         override fun length(steps: Int): Double =
             cubicLength(p0.x, p0.y, p1.x, p1.y, p2.x, p2.y, p3.x, p3.y, steps)
